static types: place in own header for testing
