+++ types/deque_list_type.toml
@@ -36,11 +36,11 @@ void getSizeType(const %1%<T, Allocator> &container, size_t& returnArg)
 handler = """
 template <typename DB, typename T0, typename T1>
 struct TypeHandler<DB, %1%<T0, T1>> {
-  using type = StaticTypes::Pair<DB,
-      StaticTypes::VarInt<DB>,
-      StaticTypes::List<DB, typename TypeHandler<DB, T0>::type>>;
+  using type = types::st::Pair<DB,
+      types::st::VarInt<DB>,
+      types::st::List<DB, typename TypeHandler<DB, T0>::type>>;
 
-  static StaticTypes::Unit<DB> getSizeType(
+  static types::st::Unit<DB> getSizeType(
       const %1%<T0, T1>& container,
       typename TypeHandler<DB, %1%<T0, T1>>::type returnArg) {
     auto tail = returnArg.write((uintptr_t)&container)
