+++ types/fb_string_type.toml
@@ -41,16 +41,16 @@ void getSizeType(const %1%<E, T, A, Storage> &container, size_t& returnArg)
 handler = """
 template <typename DB, typename T0, typename T1, typename T2, typename T3>
 struct TypeHandler<DB, %1%<T0, T1, T2, T3>> {
-  using type = StaticTypes::Pair<DB,
-      StaticTypes::VarInt<DB>,
-      StaticTypes::Pair<DB,
-        StaticTypes::VarInt<DB>,
-        StaticTypes::Pair<DB,
-          StaticTypes::VarInt<DB>,
-          StaticTypes::VarInt<DB>
+  using type = types::st::Pair<DB,
+      types::st::VarInt<DB>,
+      types::st::Pair<DB,
+        types::st::VarInt<DB>,
+        types::st::Pair<DB,
+          types::st::VarInt<DB>,
+          types::st::VarInt<DB>
       >>>;
 
-  static StaticTypes::Unit<DB> getSizeType(
+  static types::st::Unit<DB> getSizeType(
       const %1%<T0, T1, T2, T3>& container,
       typename TypeHandler<DB, %1%<T0, T1, T2, T3>>::type returnArg) {
     auto last = returnArg.write((uintptr_t)container.data())
