+++ types/multi_map_type.toml
@@ -36,12 +36,12 @@ void getSizeType(const %1%<Key,T,Compare,Allocator> &container, size_t& returnAr
 handler = """
 template <typename DB, typename T0, typename T1, typename T2, typename T3>
 struct TypeHandler<DB, %1%<T0, T1, T2, T3>> {
-  using type = StaticTypes::List<DB, StaticTypes::Pair<DB,
+  using type = types::st::List<DB, types::st::Pair<DB,
       typename TypeHandler<DB, T0>::type,
       typename TypeHandler<DB, T1>::type
     >>;
 
-  static StaticTypes::Unit<DB> getSizeType(
+  static types::st::Unit<DB> getSizeType(
       const %1%<T0, T1, T2, T3>& container,
       typename TypeHandler<DB, %1%<T0, T1, T2, T3>>::type returnArg) {
     auto tail = returnArg.write(container.size());
