+++ types/pair_type.toml
@@ -29,11 +29,11 @@ void getSizeType(const %1%<P,Q> &container, size_t& returnArg)
 handler = """
 template <typename DB, typename T0, typename T1>
 struct TypeHandler<DB, %1%<T0, T1>> {
-  using type = StaticTypes::Pair<DB,
+  using type = types::st::Pair<DB,
       typename TypeHandler<DB, T0>::type,
       typename TypeHandler<DB, T1>::type>;
 
-  static StaticTypes::Unit<DB> getSizeType(
+  static types::st::Unit<DB> getSizeType(
       const %1%<T0, T1> & container,
       typename TypeHandler<DB, %1%<T0, T1>>::type returnArg) {
     return OIInternal::getSizeType<DB>(
