+++ types/seq_type.toml
@@ -38,13 +38,13 @@ void getSizeType(const %1%<T, Allocator> &container, size_t& returnArg)
 handler = """
 template <typename DB, typename T0>
 struct TypeHandler<DB, %1% <T0>> {
-  using type = StaticTypes::Pair<
-      DB, StaticTypes::VarInt<DB>,
-      StaticTypes::Pair<
-          DB, StaticTypes::VarInt<DB>,
-          StaticTypes::List<DB, typename TypeHandler<DB, T0>::type>>>;
+  using type = types::st::Pair<
+      DB, types::st::VarInt<DB>,
+      types::st::Pair<
+          DB, types::st::VarInt<DB>,
+          types::st::List<DB, typename TypeHandler<DB, T0>::type>>>;
 
-  static StaticTypes::Unit<DB> getSizeType(
+  static types::st::Unit<DB> getSizeType(
       const %1% <T0> & container,
       typename TypeHandler<DB, %1% <T0>>::type returnArg) {
     auto tail = returnArg.write((uintptr_t)&container)
