+++ types/set_type.toml
@@ -39,11 +39,11 @@ void getSizeType(const %1%<Key, Compare, Alloc> &container, size_t& returnArg)
 handler = """
 template <typename DB, typename T0, typename T1, typename T2>
 struct TypeHandler<DB, %1% <T0, T1, T2>> {
-  using type = StaticTypes::Pair<DB,
-      StaticTypes::VarInt<DB>,
-      StaticTypes::List<DB, typename TypeHandler<DB, T0>::type>>;
+  using type = types::st::Pair<DB,
+      types::st::VarInt<DB>,
+      types::st::List<DB, typename TypeHandler<DB, T0>::type>>;
 
-  static StaticTypes::Unit<DB> getSizeType(
+  static types::st::Unit<DB> getSizeType(
       const %1%<T0, T1, T2>& container,
       typename TypeHandler<DB, %1%<T0, T1, T2>>::type returnArg) {
     constexpr size_t nodeSize = sizeof(typename %1%<T0, T1, T2>::node_type);
