+++ types/stack_container_adapter_type.toml
@@ -35,11 +35,11 @@ void getSizeType(const %1%<T, Container> &containerAdapter, size_t& returnArg)
 handler = """
 template <typename DB, typename T0, typename T1>
 struct TypeHandler<DB, %1%<T0, T1>> {
-  using type = StaticTypes::Pair<DB,
-      StaticTypes::VarInt<DB>,
+  using type = types::st::Pair<DB,
+      types::st::VarInt<DB>,
       typename TypeHandler<DB, T1>::type>;
 
-  static StaticTypes::Unit<DB> getSizeType(
+  static types::st::Unit<DB> getSizeType(
       const %1% <T0, T1> & container,
       typename TypeHandler<DB, %1% <T0, T1>>::type returnArg) {
     auto tail = returnArg.write((uintptr_t)&container);
