+++ types/std_unordered_map_type.toml
@@ -42,16 +42,16 @@ void getSizeType(const %1%<K, T, H, KE, A> &container, size_t& returnArg)
 handler = """
 template <typename DB, typename T0, typename T1, typename T2, typename T3, typename T4>
 struct TypeHandler<DB, %1%<T0, T1, T2, T3, T4>> {
-  using type = StaticTypes::Pair<DB,
-      StaticTypes::VarInt<DB>,
-      StaticTypes::Pair<DB,
-        StaticTypes::VarInt<DB>,
-        StaticTypes::List<DB, StaticTypes::Pair<DB,
+  using type = types::st::Pair<DB,
+      types::st::VarInt<DB>,
+      types::st::Pair<DB,
+        types::st::VarInt<DB>,
+        types::st::List<DB, types::st::Pair<DB,
           typename TypeHandler<DB, T0>::type,
           typename TypeHandler<DB, T1>::type
     >>>>;
 
-  static StaticTypes::Unit<DB> getSizeType(
+  static types::st::Unit<DB> getSizeType(
       const %1%<T0, T1, T2, T3, T4>& container,
       typename TypeHandler<DB, %1%<T0, T1, T2, T3, T4>>::type returnArg) {
     constexpr size_t nodeSize = sizeof(typename %1%<T0, T1, T2, T3, T4>::node_type);
